feat(search): major change in search logic from ADT (see details)
To support progress indication in the search and allow the user the option to cancel
the search at any time the scope is now persisted in a separate backend call.
The actual search requests will use the scope id and an offset to search a
specific section of the scope.

Changes:
- added new db table and ADT resource to persist the search scope from ADT
- adustments to the search API so that it still works from the SAP GUI as before
- the minimum scope size to enable parallel processing was reduced to 100