feat(parl-packaging): persists package size for parallel processing
if parallel processing was not available in the backend a package size of e.g. 1000 is
not very user friendly, as there will be no update for a while. To enable more flexibility
the package size for sequential processing will now be fixed to 100 from ADT and the 
package size for parallel processing is set in the back end.
That way the correct package size can be chosen in case parallel processing is not 
possible